\name{analyze.data.frame}
\alias{analyze.data.frame}
\title{Analyze a data...}
\usage{analyze.data.frame(obj, ...)}
\description{Analyze a \code{data.frame}}
\details{This method analyzes}
\value{an object describing}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{a \code{data.frame}}
\item{...}{additional parameters}}
