\name{analyze.formula}
\alias{analyze.formula}
\title{Analyze a Formula...}
\usage{analyze.formula(obj, data, ...)}
\description{Analyze a Formula}
\details{This function analyzes the types of data used in a formula, using a 
\code{data.frame} as reference.}
\value{a \code{data.frame.analysis} object}
\author{Matt Owen \email{mowen@iq.harvard.ed}}
\arguments{\item{obj}{a \code{formula}}
\item{data}{a \code{data.frame}}}
