\name{analyze.list}
\alias{analyze.list}
\title{Analyze a List...}
\usage{analyze.list(obj, data, ...)}
\description{Analyze a List}
\details{This function analyzes the types of data used in a formula.}
\value{a \code{data.frame.analysis} object}
\author{Matt Owen \email{mowen@iq.harvard.ed}}
\arguments{\item{obj}{a \code{list} object of analyzable data}
\item{data}{a \code{data.frame}}
\item{...}{ignored parameters}}
