\name{check.binary}
\alias{check.binary}
\title{Determine Whether Experssion has a Binary Outcome...}
\usage{check.binary(column, data)}
\description{Determine Whether Experssion has a Binary Outcome}
\details{This defines what a binary piece of data is (for Zelig)}
\value{a logical specifying whether the given expression produces binary
data as a result}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{column}{an expression representing a data column}
\item{data}{a \code{data.frame} used to evaluate \code{column}}}
