\name{check.continuous}
\alias{check.continuous}
\title{Determine Whether Experssion has a Continuous Outcome...}
\usage{check.continuous(column, data)}
\description{Determine Whether Experssion has a Continuous Outcome}
\details{This defines what a continuous piece of data is (for Zelig)}
\value{a logical specifying whether the given expression produces binary
data as a result}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{column}{an expression representing a data column}
\item{data}{a \code{data.frame} used to evaluate \code{column}}}
