\name{dvnToXml}
\alias{dvnToXml}
\title{Convert dvn File to xml...}
\usage{dvnToXml(model, dvn.file, con=stdout())}
\description{Convert \code{dvn} File to \code{xml}}
\details{Converts DVN-style documents to an XML equivalent.}
\note{This function is used}
\value{nothing}
\arguments{\item{model}{a character-string specifying the model}
\item{dvn.file}{a character-string specifying the location of a dvn-file}
\item{con}{a connection to output}}
