\name{getModelPkg}
\alias{getModelPkg}
\title{Get Package Name of a Zelig Model...}
\usage{getModelPkg(model)}
\description{Get Package Name of a Zelig Model}
\value{a character-string specifying the pacakge or NA (if the model does
not exist)}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{model}{a character-string specifying a Zelig model}}
