\name{inline.tag}
\alias{inline.tag}
\title{Create a Valid Inline XML Tag...}
\usage{inline.tag(...)}
\description{Create a Valid Inline XML Tag}
\details{This method creates a valid Inline XML tag based on the function's call object.}
\value{a block of valid XML code as a character-string}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{an arbirtaty number of parameters. The first specifies the actual
name of the tag, while all other specify its attributes. Key-value pair
style attribute, while simlpe character strings (without a parameter
specification) stand-alone within the tag.}}
