\name{multivariate.list.formula}
\alias{multivariate.list.formula}
\title{Create a Formula with a Multivariate Outcome and *System* of Explanatory Variables...}
\usage{multivariate.list.formula(outcome, explanatory, intercept, envir)}
\description{Create a Formula with a Multivariate Outcome and *System* of Explanatory Variables}
\details{...}
\value{a formula with a single outcome variable}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{outcome}{a character-vector specifying the outcome variables}
\item{explanatory}{a character-vector specifying the explanatory variables}
\item{intercept}{an integer specifying whether an intercept exists}}
