\name{print.analyze.data.frame}
\alias{print.analyze.data.frame}
\title{Print the Analysis of a data...}
\usage{\method{print}{analyze.data.frame}(x, digits=3, ...)}
\description{Print the Analysis of a \code{data.frame}}
\value{the original object (invisibly)}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a \code{analuze.data.frame} object}
\item{digits}{an integer specifying the precision of numeric variables}
\item{...}{ignored parameters}}
