\name{print.analyze.formula}
\alias{print.analyze.formula}
\title{Print the Analysis of a formula...}
\usage{\method{print}{analyze.formula}(x, digits=3, ...)}
\description{Print the Analysis of a \code{formula}}
\value{the original object (invisibly)}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a \code{analyze.formula} object}
\item{digits}{an integer specifying the precision of numeric variables}
\item{...}{ignored parameters}}
