\name{print.analyze.list}
\alias{print.analyze.list}
\title{Print the Analysis of a list...}
\usage{print.analyze.list(x, digits, ...)}
\description{Print the Analysis of a \code{list}}
\value{the original object (invisibly)}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a \code{analyze.list} analysis object}
\item{digits}{an integer specifying the precision of numeric variables}
\item{...}{ignored parameters}}
