\name{print.data.frame.analysis}
\alias{print.data.frame.analysis}
\title{Print the Analysis of a data...}
\usage{print.data.frame.analysis(x, digits, ...)}
\description{Print the Analysis of a \code{data.frame}}
\value{the original object (invisibly)}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a \code{data.frame} analysis object}
\item{digits}{an integer specifying the precision of numeric variables}
\item{...}{ignored parameters}}
