\name{print.info.atomic}
\alias{print.info.atomic}
\title{Print a Column Analysis...}
\usage{print.info.atomic(x, digits, ...)}
\description{Print a Column Analysis}
\details{Correctly format and display the analysis of an individual column.}
\value{the object (invisibly)}
\arguments{\item{x}{a \code{column.analysis} object}
\item{digits}{a numeric specifying the number of digits of precision to use
for numeric values}
\item{...}{ignored parameters}}
