\name{print.info.matrix}
\alias{print.info.matrix}
\title{Print Information about a Matrix...}
\usage{print.info.matrix(x, digits=3, ...)}
\description{Print Information about a Matrix}
\details{Correctly format and display the analysis of an individual column.}
\value{the object (invisibly)}
\arguments{\item{x}{a \code{column.analysis} object}
\item{digits}{a numeric specifying the number of digits of precision to use
for numeric values}
\item{...}{ignored parameters}}
