\name{univariate.formula}
\alias{univariate.formula}
\title{Create a Formula with a Univariate Outcome...}
\usage{univariate.formula(outcome, explanatory, intercept, envir)}
\description{Create a Formula with a Univariate Outcome}
\details{...}
\value{a formula with a single outcome variable}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{outcome}{a character-string specifying the outcome variables}
\item{explanatory}{a character-vector specifying the explanatory variables}
\item{intercept}{an integer specifying whether an intercept exists}}
