\name{ZeligDVN-package}
\title{Zelig to DVN Interface Software...}
\description{Zelig to DVN Interface Software}
\details{ZeligDVN is toolset used to analyze \code{data.frame}'s to test for 
compatability between Dataverse, Zelig and end-user input.

\tabular{ll}{
Package: \tab Zelig\cr
Version: \tab 1.0 \cr
Date: \tab 2011-09-11\cr
Depends: \tab R (>= 2.13), Zelig \cr
License: \tab GPL version 2 or newer\cr
URL: \tab http://gking.harvard.edu/zelig\cr
}}
\alias{ZeligDVN}
\docType{package}
\author{Matt Owen \email{mowen@iq.harvard.edu}
Maintainer: Matt Owen \email{monwe@iq.harvard.edu}}
\keyword{package}
\seealso{Zelig-package}
