\name{analyze}
\alias{analyze}
\title{Generic Method for Analyzing R-objects for Use with DVN...}
\usage{analyze(obj, ...)
analyze(formula, data, threshold)}
\description{Generic Method for Analyzing R-objects for Use with DVN}
\details{This method provides summary information that will be used to categorize
how and where particular types of data can be used with the DVN.}
\value{an \code{analyzed} object}
\arguments{\item{obj}{}
\item{...}{}}
