\name{print.analyze.formula}
\alias{print.analyze.formula}
\title{Print the Analysis of a formula...}
\usage{print.analyze.formula(x, digits, ...)}
\description{Print the Analysis of a \code{formula}}
\value{the original object (invisibly)}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a \code{data.frame} analysis object}
\item{digits}{an integer specifying the precision of numeric variables}
\item{...}{ignored parameters}}
