\name{print.analyze.list}
\alias{print.analyze.list}
\title{Print the Analysis of a list...}
\usage{\method{print}{analyze.list}(x, digits=3, ...)}
\description{Print the Analysis of a \code{list}}
\value{the original object (invisibly)}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a \code{analyze.list} object}
\item{digits}{an integer specifying the precision of numeric variables}
\item{...}{ignored parameters}}
