\name{print.info.atomic}
\alias{print.info.atomic}
\title{Print a Column Analysis...}
\usage{\method{print}{info.atomic}(x, ...)}
\description{Print a Column Analysis}
\details{Correctly format and display the analysis of an individual column.}
\value{the object (invisibly)}
\arguments{\item{x}{a \code{column.analysis} object}
\item{...}{ignored parameters}}
